Modularize Commands
Move the various sub commands into a hierarchical structure to
facilitate better code separation, the intention is there will be a bit
of code sharing goes on in these sub modules, so keep the concerns
separate too.
