+++ cmd/unikornctl/main.go
@@ -3,11 +3,11 @@ package main
 import (
 	"os"
 
-	"github.com/eschercloudai/unikorn/pkg/command"
+	"github.com/eschercloudai/unikorn/pkg/cmd"
 )
 
 func main() {
-	c := command.Generate()
+	c := cmd.Generate()
 
 	if err := c.Execute(); err != nil {
 		os.Exit(1)
